http_server: split window handlers into submodules and adapt SessionState callers
